pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openclawd"
version = "0.1.0"
description = "Persistent cross-project memory for Claude Code via MCP"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "lancedb>=0.6.0",
    "pyarrow>=14.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
chromadb = ["chromadb>=0.4.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.scripts]
openclawd-server = "openclawd.server:main"
openclawd-index = "scripts.index_vault:main"

[tool.setuptools.packages.find]
where = ["src"]
